Move wf1 forecast logic to forecast module, add some tests
